+++ service/collector/set.go
@@ -50,6 +50,19 @@ func NewSet(config SetConfig) (*Set, error) {
 		}
 	}
 
+	var appOperatorCollector *AppOperator
+	{
+		c := AppOperatorConfig{
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+		}
+
+		appOperatorCollector, err = NewAppOperator(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	var collectorSet *collector.Set
 	{
 		c := collector.SetConfig{
@@ -57,6 +70,11 @@ func NewSet(config SetConfig) (*Set, error) {
 				appCollector,
 			},
 			Logger: config.Logger,
+		}
+
+		if config.UniqueApp {
+			config.Logger.Log("level", "debug", "message", "app-operator collector is enabled")
+			c.Collectors = append(c.Collectors, appOperatorCollector)
 		}
 
 		collectorSet, err = collector.NewSet(c)